move device verification management to its own file
