+++ resources/res.qrc
@@ -141,37 +141,42 @@
         <file>qml/SelfVerificationCheck.qml</file>
         <file>qml/TypingIndicator.qml</file>
         <file>qml/NotificationWarning.qml</file>
-        <file>qml/emoji/EmojiPicker.qml</file>
-        <file>qml/emoji/StickerPicker.qml</file>
-        <file>qml/delegates/MessageDelegate.qml</file>
+        <file>qml/components/AdaptiveLayout.qml</file>
+        <file>qml/components/AdaptiveLayoutElement.qml</file>
+        <file>qml/components/AvatarListTile.qml</file>
+        <file>qml/components/FlatButton.qml</file>
         <file>qml/delegates/Encrypted.qml</file>
         <file>qml/delegates/FileMessage.qml</file>
         <file>qml/delegates/ImageMessage.qml</file>
+        <file>qml/delegates/MessageDelegate.qml</file>
         <file>qml/delegates/NoticeMessage.qml</file>
         <file>qml/delegates/Pill.qml</file>
         <file>qml/delegates/Placeholder.qml</file>
         <file>qml/delegates/PlayableMediaMessage.qml</file>
         <file>qml/delegates/Reply.qml</file>
         <file>qml/delegates/TextMessage.qml</file>
-        <file>qml/device-verification/Waiting.qml</file>
         <file>qml/device-verification/DeviceVerification.qml</file>
         <file>qml/device-verification/DigitVerification.qml</file>
         <file>qml/device-verification/EmojiVerification.qml</file>
-        <file>qml/device-verification/NewVerificationRequest.qml</file>
         <file>qml/device-verification/Failed.qml</file>
-		<file>qml/device-verification/Success.qml</file>
-        <file>qml/dialogs/ImagePackSettingsDialog.qml</file>
+        <file>qml/device-verification/NewVerificationRequest.qml</file>
+        <file>qml/device-verification/Success.qml</file>
+        <file>qml/device-verification/Waiting.qml</file>
         <file>qml/dialogs/ImagePackEditorDialog.qml</file>
-		<file>qml/dialogs/InputDialog.qml</file>
-		<file>qml/dialogs/InviteDialog.qml</file>
-		<file>qml/dialogs/JoinRoomDialog.qml</file>
-		<file>qml/dialogs/LogoutDialog.qml</file>
-		<file>qml/dialogs/RawMessageDialog.qml</file>
-		<file>qml/dialogs/ReadReceipts.qml</file>
-		<file>qml/dialogs/RoomDirectory.qml</file>
-		<file>qml/dialogs/RoomMembers.qml</file>
-		<file>qml/dialogs/RoomSettings.qml</file>
-		<file>qml/dialogs/UserProfile.qml</file>
+        <file>qml/dialogs/ImagePackSettingsDialog.qml</file>
+        <file>qml/dialogs/InputDialog.qml</file>
+        <file>qml/dialogs/InviteDialog.qml</file>
+        <file>qml/dialogs/JoinRoomDialog.qml</file>
+        <file>qml/dialogs/LeaveRoomDialog.qml</file>
+        <file>qml/dialogs/LogoutDialog.qml</file>
+        <file>qml/dialogs/RawMessageDialog.qml</file>
+        <file>qml/dialogs/ReadReceipts.qml</file>
+        <file>qml/dialogs/RoomDirectory.qml</file>
+        <file>qml/dialogs/RoomMembers.qml</file>
+        <file>qml/dialogs/RoomSettings.qml</file>
+        <file>qml/dialogs/UserProfile.qml</file>
+        <file>qml/emoji/EmojiPicker.qml</file>
+        <file>qml/emoji/StickerPicker.qml</file>
         <file>qml/ui/Ripple.qml</file>
         <file>qml/ui/Spinner.qml</file>
         <file>qml/ui/animations/BlinkAnimation.qml</file>
@@ -183,18 +188,6 @@
         <file>qml/voip/PlaceCall.qml</file>
         <file>qml/voip/ScreenShare.qml</file>
         <file>qml/voip/VideoCall.qml</file>
-        <file>qml/components/AdaptiveLayout.qml</file>
-        <file>qml/components/AdaptiveLayoutElement.qml</file>
-		<file>qml/components/AvatarListTile.qml</file>
-        <file>qml/components/FlatButton.qml</file>
-        <file>qml/dialogs/InviteDialog.qml</file>
-		<file>qml/dialogs/LeaveRoomDialog.qml</file>
-        <file>qml/dialogs/RawMessageDialog.qml</file>
-        <file>qml/dialogs/ReadReceipts.qml</file>
-        <file>qml/dialogs/RoomDirectory.qml</file>
-        <file>qml/dialogs/RoomMembers.qml</file>
-        <file>qml/dialogs/RoomSettings.qml</file>
-		<file>qml/dialogs/UserProfile.qml</file>
     </qresource>
     <qresource prefix="/media">
         <file>media/ring.ogg</file>
